requirements.txt
anyio==3.6.2
async-timeout==4.0.2
autopep8==2.0.2
bcrypt==3.2.2
boto3==1.26.132
botocore==1.29.132
cffi==1.15.1
cleo==1.0.0a4
click==8.1.3
crashtest==0.3.1
cryptography==40.0.2
ecdsa==0.18.0
fastapi==0.88.0
fastapi-utils==0.2.1
greenlet==2.0.2
h11==0.14.0
hiredis==2.1.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
jmespath==1.0.1
jwt==1.3.1
Mako==1.2.4
MarkupSafe==2.1.2
passlib==1.7.4
pptree==3.1
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.2
PyJWT==2.7.0
pylev==1.4.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.6
python-ulid==1.1.0
redis==4.4.0
redis-om==0.0.20
rsa==4.9
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.22.0
tomli==2.0.1
types-redis==4.3.21.7
types-six==1.16.21.4
typing-extensions==4.4.0
urllib3==1.26.15
uvicorn==0.20.0
yappi==1.4.0
zipp==3.15.0
